pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "imgfind"
version = "0.1.0"
description = "Search images with specified color"
authors = ["h5vx <h5v@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
scikit-image = "^0.17.2"
opencv-python = "^4.4.0"
python-magic = "^0.4.18"
attrs = "^20.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6"
black = "^20.8b1"
isort = "^5.5.3"
pip-tools = "^5.3.1"
